+++ packages/component/src/Activity/Avatar.js
@@ -1,50 +1,14 @@
-import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import connectToWebChat from '../connectToWebChat';
-import CroppedImage from '../Utils/CroppedImage';
-import useAvatarForBot from '../hooks/useAvatarForBot';
-import useAvatarForUser from '../hooks/useAvatarForUser';
-import useStyleSet from '../hooks/useStyleSet';
-
-const connectAvatar = (...selectors) =>
-  connectToWebChat(
-    (
-      {
-        styleSet: {
-          options: { botAvatarImage, botAvatarInitials, userAvatarImage, userAvatarInitials }
-        }
-      },
-      { fromUser }
-    ) => ({
-      avatarImage: fromUser ? userAvatarImage : botAvatarImage,
-      avatarInitials: fromUser ? userAvatarInitials : botAvatarInitials
-    }),
-    ...selectors
-  );
-
-// TODO: [P2] Consider memoizing "style={ backgroundImage }" in our upstreamers
-//       We have 2 different upstreamers <CarouselFilmStrip> and <StackedLayout>
+import { DefaultAvatar } from '../Middleware/Avatar/createCoreMiddleware';
 
 const Avatar = ({ 'aria-hidden': ariaHidden, className, fromUser }) => {
-  const [botAvatar] = useAvatarForBot();
-  const [userAvatar] = useAvatarForUser();
-  const [{ avatar: avatarStyleSet }] = useStyleSet();
-
-  const { image, initials } = fromUser ? userAvatar : botAvatar;
-
-  return (
-    !!(image || initials) && (
-      <div
-        aria-hidden={ariaHidden}
-        className={classNames(avatarStyleSet + '', { 'from-user': fromUser }, className + '')}
-      >
-        {initials}
-        {!!image && <CroppedImage alt="" className="image" height="100%" src={image} width="100%" />}
-      </div>
-    )
+  console.warn(
+    'botframework-webchat: <Avatar> component is deprecated and will be removed on or after 2022-02-25. Please use `useRenderAvatar` hook instead.'
   );
+
+  return <DefaultAvatar aria-hidden={ariaHidden} className={className} fromUser={fromUser} />;
 };
 
 Avatar.defaultProps = {
@@ -60,5 +24,3 @@ Avatar.propTypes = {
 };
 
 export default Avatar;
-
-export { connectAvatar };