+++ packages/component/src/Styles/StyleSet/StackedLayout.js
@@ -17,51 +17,51 @@ export default function createStackedLayoutStyle({ bubbleMaxWidth, bubbleMinWidt
     '&:not(.webchat__stacked_extra_right_indent)': {
       marginRight: paddingRegular
     },
-    '&:not(.webchat__stacked--rtl)': {
-      '&:not(.from-user)': {
-        '&.webchat__stacked_indented_content > .avatar': {
+    '&:not(.webchat__stackedLayout--rtl)': {
+      '&:not(.webchat__stackedLayout--fromUser)': {
+        '&.webchat__stacked_indented_content > .webchat__stackedLayout__avatar': {
           marginRight: paddingRegular
         },
 
-        '& > .content > .webchat__stacked_item_indented': {
+        '& > .webchat__stackedLayout__content > .webchat__stacked_item_indented': {
           marginLeft: paddingRegular
         }
       },
 
-      '&.from-user': {
-        '&.webchat__stacked_indented_content > .avatar': {
+      '&.webchat__stackedLayout--fromUser': {
+        '&.webchat__stacked_indented_content > .webchat__stackedLayout__avatar': {
           marginLeft: paddingRegular
         },
 
-        '& > .content > .webchat__stacked_item_indented': {
+        '& > .webchat__stackedLayout__content > .webchat__stacked_item_indented': {
           marginRight: paddingRegular
         }
       }
     },
 
-    '&.webchat__stacked--rtl': {
-      '&:not(.from-user)': {
-        '&.webchat__stacked_indented_content > .avatar': {
+    '&.webchat__stackedLayout--rtl': {
+      '&:not(.webchat__stackedLayout--fromUser)': {
+        '&.webchat__stacked_indented_content > .webchat__stackedLayout__avatar': {
           marginLeft: paddingRegular
         },
 
-        '& > .content > .webchat__stacked_item_indented': {
+        '& > .webchat__stackedLayout__content > .webchat__stacked_item_indented': {
           marginRight: paddingRegular
         }
       },
 
-      '&.from-user': {
-        '&.webchat__stacked_indented_content > .avatar': {
+      '&.webchat__stackedLayout--fromUser': {
+        '&.webchat__stacked_indented_content > .webchat__stackedLayout__avatar': {
           marginRight: paddingRegular
         },
 
-        '& > .content > .webchat__stacked_item_indented': {
+        '& > .webchat__stackedLayout__content > .webchat__stacked_item_indented': {
           marginLeft: paddingRegular
         }
       }
     },
 
-    '& > .content': {
+    '& > .webchat__stackedLayout__content': {
       '& > .webchat__row': {
         '& > .bubble, & > .timestamp': {
           maxWidth: bubbleMaxWidth