+++ packages/component/src/Styles/defaultStyleOptions.js
@@ -26,6 +26,7 @@ const DEFAULT_OPTIONS = {
   primaryFont: fontFamily(['Calibri', 'Helvetica Neue', 'Arial', 'sans-serif']),
 
   // Avatar
+  avatarBorderRadius: '50%',
   avatarSize: 40,
   botAvatarBackgroundColor: undefined, // defaults to accent color
   botAvatarImage: '',